refactor util out of cli module
