+++ jaxgmg/cli/mazegen.py
@@ -7,7 +7,7 @@ import jax.numpy as jnp
 import einops
 
 from jaxgmg.procgen import maze_generation
-from jaxgmg.cli import util
+from jaxgmg import util
 
 
 def tree(
