+++ jaxgmg/cli/parse.py
@@ -12,7 +12,7 @@ from jaxgmg.environments import monster_world
 from jaxgmg.environments import lava_land
 from jaxgmg.environments import follow_me
 
-from jaxgmg.cli import util
+from jaxgmg import util
 
 
 def corner(
