+++ jaxgmg/cli/play.py
@@ -15,7 +15,7 @@ from jaxgmg.environments import keys_and_chests
 from jaxgmg.environments import monster_world
 from jaxgmg.environments import lava_land
 from jaxgmg.environments import follow_me
-from jaxgmg.cli import util
+from jaxgmg import util
 
 
 # # #
