+++ jaxgmg/util.py
@@ -1,56 +1,21 @@
 """
-Utilities supporting jaxgmg CLI applications.
+Utilities for transforming data or rendering it to strings/stdout/disk/wandb
+in support of jaxgmg CLI application and training scripts.
 """
 
 import functools
+import os
+import datetime
 import numpy as np
 import jax.numpy as jnp
 import einops
 import PIL.Image as pillow
-import typer
 import wandb
-
-
-def make_typer_app(name, help, subcommands, **kwargs):
-    """
-    Transform a list of functions into a typer application.
-    """
-    app = typer.Typer(name=name, help=help, **kwargs)
-    for subcommand in subcommands:
-        app.command()(subcommand)
-    return app
-
-
-def print_config(config: dict):
-    """
-    Dump a dictionary's contents to stdout.
-    """
-    for key, value in config.items():
-        print(f"  {key:30s}: {value}")
-
-
-def print_histogram(data, bins=10, range=None, width=40):
-    """
-    Bin and count a sequence of values and print them to stdout as an ASCII
-    histogram.
-    """
-    hist, bin_edges = jnp.histogram(data, bins=bins, range=range)
-    norm_counts = hist / hist.max()
-    for count, lo, hi in zip(norm_counts, bin_edges, bin_edges[1:]):
-        print(f"  {lo:.2f} to {hi:.2f} | {int(count * width + 1) * '*'}")
-
-
-def print_legend(legend, colormap=None):
-    """
-    Render a mapping from colors to values to stdout.
-    """
-    print("legend:")
-    for value, name in legend.items():
-        print(img2str(jnp.full((2,2), value), colormap=colormap,), name)
+import json
 
 
 # # # 
-# Image and animation rendering
+# Rendering data / images / plots to strings
 
 
 def img2str(im, colormap=None):
@@ -88,6 +53,100 @@ def img2str(im, colormap=None):
         "".join([_switch_color(fg, bg) + "▀" for fg, bg in row])
         for row in im
     ]) + _reset
+
+
+def dict2str(dct):
+    def dict2lines(dct, depth):
+        for key, value in dct.items():
+            if isinstance(value, dict):
+                yield (depth, key, '')
+                yield from dict2lines(value, depth+1)
+            else:
+                yield (depth, key, value)
+    return '\n'.join([
+        '  '*depth + (key + ':').ljust(42-2*depth) + str(value)
+        for depth, key, value in dict2lines(dct, 1)
+    ])
+
+
+# # # 
+# Rendering data / images / plots to stdout
+
+
+def print_config(config: dict):
+    """
+    Dump a dictionary's contents to stdout.
+    """
+    for key, value in config.items():
+        print(f"  {key:30s}: {value}")
+
+
+def print_histogram(data, bins=10, range=None, width=40):
+    """
+    Bin and count a sequence of values and print them to stdout as an ASCII
+    histogram.
+    """
+    hist, bin_edges = jnp.histogram(data, bins=bins, range=range)
+    norm_counts = hist / hist.max()
+    for count, lo, hi in zip(norm_counts, bin_edges, bin_edges[1:]):
+        print(f"  {lo:.2f} to {hi:.2f} | {int(count * width + 1) * '*'}")
+
+
+def print_legend(legend, colormap=None):
+    """
+    Render a mapping from colors to values to stdout.
+    """
+    print("legend:")
+    for value, name in legend.items():
+        print(img2str(jnp.full((2,2), value), colormap=colormap,), name)
+
+
+def print_img(im, colormap=None):
+    """
+    Render a small image to stdout using unicode half-block characters to
+    represent pairs of pixels.
+    """
+    print(img2str, colormap=colormap)
+
+
+# # # 
+# Saving images or animations to disk
+
+
+def save_image(
+    image,
+    path,
+    upscale=1,
+):
+    """
+    Take a (height, width, rgb) matrix and save it as a png.
+    
+    Parameters:
+
+    * image: float[h, w, rgb]
+            The image. Axes represent the image data. Each point should be a
+            float between 0 and 1.
+    * path : str
+            Where to save the image.
+    * upscale : int (>=1, default is 1)
+            Width/height of pixel representation of each matrix entry.
+    """
+    image = np.asarray(image)
+    H, W, C = image.shape
+    assert C == 3, f"Wrong number of channels for GIF (C={C}!=3)"
+        
+    # preprocess image data
+    image_u8 = (image * 255).astype(np.uint8)
+    image_u8_upscaled = einops.repeat(
+        image_u8,
+        'h w rgb -> (h sh) (w sw) rgb',
+        sh=upscale,
+        sw=upscale,
+    )
+    # PIL images for each frame
+    img = pillow.fromarray(image_u8_upscaled)
+    # save!
+    img.save(path)
 
 
 def save_gif(
@@ -139,44 +198,66 @@ def save_gif(
     )
 
 
-def save_image(
-    image,
-    path,
-    upscale=1,
-):
-    """
-    Take a (height, width, rgb) matrix and save it as a png.
-    
-    Parameters:
-
-    * image: float[h, w, rgb]
-            The image. Axes represent the image data. Each point should be a
-            float between 0 and 1.
-    * path : str
-            Where to save the image.
-    * upscale : int (>=1, default is 1)
-            Width/height of pixel representation of each matrix entry.
-    """
-    image = np.asarray(image)
-    H, W, C = image.shape
-    assert C == 3, f"Wrong number of channels for GIF (C={C}!=3)"
-        
-    # preprocess image data
-    image_u8 = (image * 255).astype(np.uint8)
-    image_u8_upscaled = einops.repeat(
-        image_u8,
-        'h w rgb -> (h sh) (w sw) rgb',
-        sh=upscale,
-        sw=upscale,
-    )
-    # PIL images for each frame
-    img = pillow.fromarray(image_u8_upscaled)
-    # save!
-    img.save(path)
+def save_json(dct, path):
+    with open(path, 'w') as outfile:
+        json.dump(dct, outfile, indent=True)
 
 
 # # # 
-# WANDB WRAPPERS AND FORMATTING FUNCTIONS
+# wandb wrappers and formatting functions
+
+
+def wandb_img(image):
+    """
+    Format a gif as a video for wandb.
+
+    Parameters:
+
+    * image : float[h, w, rgb]
+            RGB floats each channel in range [0,1].
+
+    Returns:
+
+    * image : wandb.Image (contains uint8[h, w, c]).
+            RGB Image including this data in the required format.
+
+    """
+    return wandb.Image(
+        np.asarray(
+            255 * image,
+            dtype=np.uint8,
+        ),
+    )
+
+
+def wandb_gif(frames, fps=12):
+    """
+    Format a gif as a video for wandb.
+
+    Parameters:
+
+    * frames : float[t h w c]
+            RGB floats each channel in range [0,1].
+    * fps : int = 12
+            Frames per second for the wandb video.
+
+    Returns:
+
+    * video : wandb.Video (contains uint8[t c h w]).
+            RGB video including this data in the required format.
+
+    """
+    return wandb.Video(
+        np.asarray(
+            255 * einops.rearrange(frames, 't h w c -> t c h w'),
+            dtype=np.uint8,
+        ),
+        fps=fps,
+    )
+
+
+# # # 
+# Training run file/wandb management
 
 
 def wandb_run(f):
@@ -253,8 +334,36 @@ def wandb_run(f):
     return g
 
 
+class RunFilesManager:
+    def __init__(self, root_path="out/"):
+        now = datetime.datetime.now().strftime('%Y-%m-%d_%H-%M-%S')
+        self.path = os.path.abspath(os.path.join(root_path, f'run_{now}'))
+        os.makedirs(self.path, exist_ok=False)
+
+
+    def get_path(self, suffix):
+        path = os.path.join(self.path, suffix)
+        os.makedirs(os.path.dirname(path), exist_ok=True)
+        return path
+
+
 # # # 
-# COLORMAPS
+# Transforming dictionaries
+
+
+def flatten_dict(nested_dict, separator='/'):
+    merged_dict = {}
+    for key, inner_dict in nested_dict.items():
+        if isinstance(inner_dict, dict):
+            for inner_key, value in flatten_dict(inner_dict).items():
+                merged_dict[key + separator + inner_key] = value
+        else:
+            merged_dict[key] = inner_dict
+    return merged_dict
+
+
+# # # 
+# Colormaps
 
 
 def viridis(x):
